+++ src/Components/Stacks.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Nav from "../Nav/Nav";
+import Nav from "./Nav";
 
 const Stacks = () => {
   return (
