+++ src/main.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import App from './App.jsx';
-import Technologies from './Technologies/Technologies.jsx';
-import Stacks from './Stacks/Stacks.jsx';
+import Technologies from './Components/Technologies.jsx';
+import Stacks from './Components/Stacks.jsx';
 import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
